recover lock
